Start factoring out tracing index page into its own folder and component files
