+++ packages/playground/src/components/traces/TracesList/index.ts
@@ -1,0 +1,2 @@
+export { TracesList } from "./TracesList";
+export { TracesListErrorBoundary } from "./ErrorBoundary";
