+++ packages/playground/src/routes/__root.tsx
@@ -35,6 +35,15 @@ export const Route = createRootRouteWithContext<{
       },
     };
   },
+
+  pendingComponent: () => {
+    return (
+      <div className="flex flex-col items-center justify-center h-full p-4">
+        <h2 className="mb-2 text-lg font-medium">Loading API Spec</h2>
+      </div>
+    );
+  },
+
   staleTime: 10 * 60 * 1000,
 
   onError: (error) => {
@@ -70,7 +79,7 @@ function RootComponent() {
         <Outlet />
       </div>
       {/*  Commented out because they're annoying but leaving them here in case you need them */}
-      {/* <TanStackRouterDevtools position="bottom-right" /> */}
+      <TanStackRouterDevtools position="bottom-right" />
       {/* <ReactQueryDevtools /> */}
     </div>
   );
